feat: 단어장이 존재한다면 특정 조건을 통해 정답을 추출할 수 있다
