+++ src/test/java/wordle/domain/WordListFileReader.java
@@ -1,0 +1,33 @@
+package wordle.domain;
+
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+
+class WordListFileReader implements WordListReader {
+    private static final String FILE_PATH = "src/main/resources/words.txt";
+    private static final WordList wordList = initializeWordList();
+
+    private static WordList initializeWordList() {
+        try {
+            final Path path = Paths.get(FILE_PATH);
+            return new WordList(Files.readAllLines(path)
+                    .stream()
+                    .map(Word::new)
+                    .toList());
+
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    public Answer read(final Selector selector) {
+        return new Answer(wordList.select(selector).getWord());
+    }
+
+    @Override
+    public WordList read() {
+        return null;
+    }
+}
