+++ src/test/java/wordle/domain/WordListReader.java
@@ -1,0 +1,7 @@
+package wordle.domain;
+
+interface WordListReader {
+    Answer read(final Selector selector);
+
+    WordList read();
+}
