+++ src/test/java/wordle/domain/WordListReaderTest.java
@@ -10,11 +10,11 @@ public class WordListReaderTest {
 
     @Test
     void 단어장이_존재한다면_특정_조건을_통해_정답을_추출할_수_있다() {
-        WordListReader wordListReader = new WordListFileReader();
+        WordListReader wordListReader = new TestWordListReader();
         WordList wordList = wordListReader.read();
 
         final Word actual = wordList.select(new TestSelector());
-        assertEquals(new Answer("apple"), actual);
+        assertEquals(new Word("apple"), actual);
     }
 
     class TestSelector implements Selector {
@@ -22,6 +22,19 @@ public class WordListReaderTest {
         @Override
         public Word select(final List<Word> wordList) {
             return wordList.getFirst();
+        }
+    }
+
+    class TestWordListReader implements WordListReader {
+
+        @Override
+        public Answer read(final Selector selector) {
+            throw new RuntimeException();
+        }
+
+        @Override
+        public WordList read() {
+            return new WordList(List.of(new Word("apple")));
         }
     }
 }
